+++ MatrixPilot/gain_variables.h
@@ -20,15 +20,15 @@
 
 #if ((SERIAL_OUTPUT_FORMAT == SERIAL_MAVLINK) || ( GAINS_VARIABLE == 1 ))
    // MAVLINK, QGROUND CONTROL (Ground Control Station) can change these variables 
-	extern int rollkp;
-	extern int rollkd;
-	extern int yawkpail;
-	extern int yawkdail;
+	extern int16_t rollkp;
+	extern int16_t rollkd;
+	extern int16_t yawkpail;
+	extern int16_t yawkdail;
 #else
-	extern const int rollkp;
-	extern const int rollkd;
-	extern const int yawkpail;
-	extern const int yawkdail;
+	extern const int16_t rollkp;
+	extern const int16_t rollkd;
+	extern const int16_t yawkpail;
+	extern const int16_t yawkdail;
 #endif
 //#define AILERON_BOOST						1.0
 
@@ -39,15 +39,15 @@
 // AILERON_ELEV_MIX is the degree of elevator adjustment for aileron
 // ELEVATOR_BOOST is the additional gain multiplier for the manually commanded elevator deflection
 #if ((SERIAL_OUTPUT_FORMAT == SERIAL_MAVLINK) || ( GAINS_VARIABLE == 1 ))
-	extern int pitchgain;
-	extern int pitchkd;
-	extern int rudderElevMixGain;
-	extern int rollElevMixGain;
+	extern int16_t pitchgain;
+	extern int16_t pitchkd;
+	extern int16_t rudderElevMixGain;
+	extern int16_t rollElevMixGain;
 #else
-	extern const int pitchgain;
-	extern const int pitchkd;
-	extern const int rudderElevMixGain;
-	extern const int rollElevMixGain;
+	extern const int16_t pitchgain;
+	extern const int16_t pitchkd;
+	extern const int16_t rudderElevMixGain;
+	extern const int16_t rollElevMixGain;
 #endif
 //#define ELEVATOR_BOOST						0.5
 
@@ -64,17 +64,17 @@
 // in stabilized or waypoint mode.  This mainly helps aileron-initiated turning while in stabilized.
 // RUDDER_BOOST is the additional gain multiplier for the manually commanded rudder deflection
 #if ((SERIAL_OUTPUT_FORMAT == SERIAL_MAVLINK) || ( GAINS_VARIABLE == 1 ))
-	extern int yawkprud;
-	extern int yawkdrud;
-	extern int rollkprud;
-	extern int rollkdrud;
+	extern int16_t yawkprud;
+	extern int16_t yawkdrud;
+	extern int16_t rollkprud;
+	extern int16_t rollkdrud;
 #else
-	extern const int yawkprud;
-	extern const int yawkdrud;
-	extern const int rollkprud;
-	extern const int rollkdrud;
+	extern const int16_t yawkprud;
+	extern const int16_t yawkdrud;
+	extern const int16_t rollkprud;
+	extern const int16_t rollkdrud;
 #endif
-//extern int MANUAL_AILERON_RUDDER_MIX			0.20
+//extern int16_t MANUAL_AILERON_RUDDER_MIX			0.20
 //#define RUDDER_BOOST						1.0
 
 // Gains for Hovering
@@ -92,19 +92,19 @@
 //                            value is proportionally scaled down.
 
 #if ((SERIAL_OUTPUT_FORMAT == SERIAL_MAVLINK) || ( GAINS_VARIABLE == 1 ))
-	extern int hoverrollkp;
-	extern int hoverrollkd;
-	extern int hoverpitchgain;
-	extern int hoverpitchkd;
-	extern int hoveryawkp;
-	extern int hoveryawkd;
+	extern int16_t hoverrollkp;
+	extern int16_t hoverrollkd;
+	extern int16_t hoverpitchgain;
+	extern int16_t hoverpitchkd;
+	extern int16_t hoveryawkp;
+	extern int16_t hoveryawkd;
 #else
-	extern const int hoverrollkp;
-	extern const int hoverrollkd;
-	extern const int hoverpitchgain;
-	extern const int hoverpitchkd;
-	extern const int hoveryawkp;
-	extern const int hoveryawkd;
+	extern const int16_t hoverrollkp;
+	extern const int16_t hoverrollkd;
+	extern const int16_t hoverpitchgain;
+	extern const int16_t hoverpitchkd;
+	extern const int16_t hoveryawkp;
+	extern const int16_t hoveryawkd;
 #endif
 
 //#define HOVER_PITCH_OFFSET					0.0		// + leans towards top, - leans towards bottom
@@ -115,22 +115,22 @@
 
 /*
 // servo throw can be more than 3 turns - 1080 degrees - so use integers rather than char
-const int tan_pitch_in_stabilized_mode;
-const int yaw_in_stabilized_mode;
+const int16_t tan_pitch_in_stabilized_mode;
+const int16_t yaw_in_stabilized_mode;
 
-const int pitch_offset_centred;
-const int yaw_offset_centred;
+const int16_t pitch_offset_centred;
+const int16_t yaw_offset_centred;
 
-const int pitch_servo_max;
-const int pitch_servo_min;
-const int yaw_servo_max;
-const int yaw_servo_min;
+const int16_t pitch_servo_max;
+const int16_t pitch_servo_min;
+const int16_t yaw_servo_max;
+const int16_t yaw_servo_min;
 
 // servo_ratios are used to convert degrees of rotation into servo pulse code lengths
 // This code is configured for the full throw of the servo to be achieved by a range of
 // 2000 units being sent to udb_pwOut. (i.e. min 2000, centered 3000, max 4000 )
-const int pitch_servo_ratio;
-const int yaw_servo_ratio;
+const int16_t pitch_servo_ratio;
+const int16_t yaw_servo_ratio;
 
 */
